pyproject.toml
[tool.poetry]
name = "profetrack"
version = "0.1.0"
description = ""
authors = ["ecaf1 <edeilsonazevedo22@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
mysql-connector-python = "^8.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
